solution/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic[email]==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
lark>=1.1.9
redis>=5.0.1
fastapi-cache2>=0.2.1
pytest